http - postProcess() to make Date objects
JSON has this unfortunate inability to handle Date objects. So when you get an object that had dates before, they're now strings. But jugglingdb knows the schema of objects in question, so it can do something about this!

installPostProcessor() is called for a model on initial setup - it looks for all fields with the type Date, and creates a postProcess() method for that model which turns all of those strings into dates.

We call postProcess() with the results of find, all, and save/update. :0)